+++ test/acceptance/export/shapefile.js
@@ -194,10 +194,37 @@ test('mixed type geometry', function(done){
         encoding: 'binary',
         method: 'GET'
     },{ }, function(res){
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
         assert.equal(res.statusCode, 400, res.statusCode + ': ' +res.body);
         var parsedBody = JSON.parse(res.body);
         var expectedBody = {"error":["ERROR 1: Attempt to write non-point (LINESTRING) geometry to point shapefile."]}
         assert.deepEqual(parsedBody, expectedBody);
+        done();
+    });
+});
+
+test('skipfields controls fields included in SHP output', function(done){
+    var query = querystring.stringify({
+        q: "SELECT 111 as skipme, 222 as keepme, 'POINT(0 0)'::geometry as g",
+        format: 'shp',
+        skipfields: 'skipme',
+        filename: 'myshape'
+      });
+    assert.response(app, {
+        url: '/api/v1/sql?' + query,
+        headers: {host: 'vizzuality.cartodb.com'},
+        encoding: 'binary',
+        method: 'GET'
+    },{ }, function(res){
+        assert.equal(res.statusCode, 200, res.body);
+        var tmpfile = '/tmp/myshape.zip';
+        var err = fs.writeFileSync(tmpfile, res.body, 'binary');
+        if (err) { done(err); return }
+        var zf = new zipfile.ZipFile(tmpfile);
+        var buffer = zf.readFileSync('myshape.dbf');
+        fs.unlinkSync(tmpfile);
+        var strings = buffer.toString();
+        assert.ok(!/skipme/.exec(strings), "Could not skip 'skipme' field:\n" + strings);
         done();
     });
 });