+++ NEWS.md
@@ -1,5 +1,6 @@
 1.3.5 (DD/MM/YY)
 -----
+* Fix skipfields use with SHP output format (#81)
 
 1.3.4 (21/01/13)
 -----